refactor: 폴더 경로 변경
