+++ src/pages/ProductDetailPage.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
-import { getProductItem } from '../api';
+import { getProductItem } from '../store/api';
 import { StyledImageBox, StyledImg } from '../components/common/Styled';
 import { COLORS } from '../styles/theme';
